Feat - 프로젝트에 모달 환경설정 적용 (#6)
