+++ index.js
@@ -7,14 +7,19 @@ import {Store} from './store/Store';
 import {NavigationContainer} from '@react-navigation/native';
 import {light} from './src/style/theme';
 import {ThemeProvider} from 'styled-components/native'; // 사용하는 테마 라이브러리에 따라 다를 수 있습니다.
+import {ModalProvider} from './src/Modal/ModalProvider';
+import CustomModal from './src/Modal/CustomModal';
 
 export default function Main() {
   return (
     <StoreProvider store={Store}>
       <ThemeProvider theme={light}>
-        <NavigationContainer>
-          <App />
-        </NavigationContainer>
+        <ModalProvider>
+          <NavigationContainer>
+            <CustomModal />
+            <App />
+          </NavigationContainer>
+        </ModalProvider>
       </ThemeProvider>
     </StoreProvider>
   );
