+++ package.json
@@ -19,17 +19,20 @@
     "@types/react-native-vector-icons": "^6.4.15",
     "react": "18.2.0",
     "react-native": "^0.72.7",
+    "react-native-calendars": "^1.1302.0",
+    "react-native-debugger-open": "^0.4.3",
     "react-native-animated-spinkit": "^1.5.2",
-    "react-native-paper": "^5.10.6",
     "react-native-safe-area-context": "^4.7.2",
     "react-native-screens": "^3.25.0",
     "react-native-simple-modal": "^9.0.1",
     "react-native-vector-icons": "^10.0.0",
     "react-redux": "^8.1.2",
     "redux": "^4.2.1",
+    "redux-devtools-extension": "^2.13.9",
     "redux-logger": "^3.0.6",
     "styled-components": "^6.1.1",
-    "zustand": "^4.4.6"
+    "zustand": "^4.4.6",
+    "rn-emoji-keyboard": "^1.5.1"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
